Add files via upload
This script contains four main functions that serve as the analysis layer for web scraping TikTok data associated with certain hashtags. The script uses the Zeeschuimer module to scrape data from user profiles and hashtags. It then extracts unique video IDs from the downloaded ndjson files and generates a TikTok link for each video. The resultant dataframe has labels for hashtag videos and user profiles and includes their associated video links.

Function 1: auto_data_file_importer_hashtag
This function imports a data file in ndjson format associated with a given hashtag. It then converts this data into a pandas DataFrame, and saves the DataFrame as a CSV file. This function is useful for importing and preprocessing data files related to specific hashtags for subsequent analysis.

Function 2: extract_usernames_and_ids_hashtag
This function takes a DataFrame and a hashtag name as inputs and extracts the usernames and video IDs associated with the given hashtag. It also counts the occurrences of each username to determine if the same users are posting multiple videos under the same hashtag. The function returns a tuple consisting of a list of username and video ID pairs, and a dictionary with username occurrences. This function is helpful for extracting relevant information from a DataFrame related to a specific hashtag and analyzing username patterns.

Function 3: tiktok_link_generator_hashtag_generalized
This function generates TikTok video links given a list of user ID and video ID pairs. It constructs the TikTok video link using the base URL, username with "@" sign, video ID, and additional parameters for each pair, and returns a list of these generated TikTok video links.

Function 4: save_videolinks_to_csv_hashtag
This function saves a list of video links to a CSV file. It first converts the list of video links into a pandas DataFrame, then constructs the filename for the CSV file using the hashtag name and additional information, and finally saves the DataFrame as a CSV file in a specified location.

In summary, the script makes it convenient to scrape data associated with specific hashtags from TikTok, extract relevant information such as usernames and video IDs, generate links to the corresponding TikTok videos, and save these links in a CSV file for further analysis.